Move set_widget_selected_index_state out of the plugins-gated module
The helper lived in plugin_dispatch.rs (compiled only with the `plugins`
feature) but is also called from widget_runtime.rs, which builds even under
`--no-default-features` — breaking the "check (no plugins)" CI job with an
E0599. Move it to widget_runtime.rs (always compiled); the plugins-gated
SetSelectedIndex mutation can still reach it since both are `impl Editor`.
